+++ NEWS.md
@@ -1,5 +1,6 @@
 # patRoon 1.0.3
 * `componentsNT`: include algorithm data returned by `nontarget::homol.search` in `homol` slot (suggested by Vittorio Albergamo)
+* `convertMSFiles()`: prevent error when no input files are found (issue #14).
 
 
 # patRoon 1.0.2