+++ NEWS.md
@@ -6,6 +6,7 @@
     * recognize that Waters files are directories
 * minor fixes for subsetting XCMS `features` objects
 * Fixed: `generateCompoundsMetFrag()`: compound names could be sometimes be interpreted as dates (reported by Corey Griffith)
+* Fixed: on very rare cases empty peaklists could be present after averaging.
 
 
 # patRoon 1.0.2