+++ NEWS.md
@@ -10,6 +10,7 @@
 * Fixed: on very rare cases empty peaklists could be present after averaging.
 * Fixed: `SIRIUS` annotation didn't use set adduct but used default instead
 * `SIRIUS` results are better handled if choosen adduct is not `[M+H]+` or `[M+H]+`
+* More fixes for loading `data.table` objects properly from cache.
 
 
 # patRoon 1.0.2